chore: initial for error handling plugin
