+++ src/app.vue
@@ -3,7 +3,6 @@
     <v-app>
       <nuxt-layout>
         <nuxt-loading-indicator />
-
         <nuxt-page />
       </nuxt-layout>
     </v-app>
