+++ src/error.vue
@@ -1,18 +1,13 @@
 <template>
-  <v-theme-provider>
-    <v-app>
-      <v-layout class="error-page flex-column" full-height>
-        <div class="text-center">
-          <h2 class="mt-10">{{ error?.statusCode }}</h2>
-          <h2 class="mt-4">{{ error }}</h2>
-        </div>
-      </v-layout>
-    </v-app>
-  </v-theme-provider>
+  <div class="text-center">
+    <h2 class="mt-10">{{ error?.statusCode }}</h2>
+    <h2 class="mt-4">{{ error }}</h2>
+  </div>
 </template>
 
 <script lang="ts" setup>
 const error = useError()
+console.log(error)
 
 useHead({
   titleTemplate: (titleChunk) => {
